feat(backend): add Ollama AI config, migrations V110-V114, fix controller compile errors
- app.ai.ollama-url/model (env-configurable OLLAMA_URL/AI_MODEL) for the
  AiAssistantService fallback engine backed by local Ollama
- migrations: ai_chat_conversations, ai_visit_notes, recurring_donations,
  admin_requests, missing_entity_tables
- AdminStatsController doubleValue() cast fix; BenchmarkController uses real
  repositories/rates (StatutAme.ACTIF, non-blank reports) with rewritten tests

+++ backend/src/main/java/com/discipolat/modules/admin/api/AdminStatsController.java
@@ -39,7 +39,7 @@ public class AdminStatsController {
         long newConverts = soulRepository.countByTypeDisciple(TypeDisciple.NOUVEAU_CONVERTI);
 
         Map<String, Object> kpi = dashboardService.getKPI(null, null, null, null);
-        double attendanceRate = (Number) kpi.getOrDefault("tauxPresenceGlobal", 0.0);
+        double attendanceRate = ((Number) kpi.getOrDefault("tauxPresenceGlobal", 0.0)).doubleValue();
         double growthRate = totalMembers > 0
                 ? Math.round((double) newConverts / totalMembers * 1000.0) / 10.0
                 : 0.0;
